+++ mitra_api/src/mastodon_api/instance/types.rs
@@ -410,6 +410,7 @@ pub struct InstanceInfoV2 {
     authentication_methods: Vec<String>,
     login_message: String,
     new_accounts_read_only: bool,
+    like_emoji: String,
     blockchains: Vec<BlockchainInfo>,
     ipfs_gateway_url: Option<String>,
 
@@ -495,6 +496,7 @@ impl InstanceInfoV2 {
             login_message: config.login_message.clone(),
             new_accounts_read_only:
                 matches!(config.registration.default_role, DefaultRole::ReadOnlyUser),
+            like_emoji: dynamic_config.like_emoji,
             blockchains: config.blockchains().iter()
                 .map(BlockchainInfo::from)
                 .collect(),