+++ mitra_cli/src/commands/config.rs
@@ -14,6 +14,7 @@ use mitra_models::{
             FEDERATED_TIMELINE_RESTRICTED,
             FILTER_BLOCKLIST_PUBLIC,
             FILTER_KEYWORDS,
+            LIKE_EMOJI,
         },
         queries::{
             set_internal_property,
@@ -32,6 +33,9 @@ enum ParameterName {
     /// Keywords for reject-keywords filter action (an array of strings, example: ["foo", "bar"])
     #[clap(name = FILTER_KEYWORDS)]
     FilterKeywords,
+    /// Emoji that is used to represent the "like" reaction ("thumbs_up" or "heart", default: "thumbs_up")
+    #[clap(name = LIKE_EMOJI)]
+    LikeEmoji,
 }
 
 impl ParameterName {
@@ -40,6 +44,7 @@ impl ParameterName {
             Self::FederatedTimelineRestricted => FEDERATED_TIMELINE_RESTRICTED,
             Self::FilterBlocklistPublic => FILTER_BLOCKLIST_PUBLIC,
             Self::FilterKeywords => FILTER_KEYWORDS,
+            Self::LikeEmoji => LIKE_EMOJI,
         };
         assert!(EDITABLE_PROPERTIES.contains(&name_str));
         name_str
@@ -82,7 +87,11 @@ impl UpdateConfig {
         db_pool: &DatabaseConnectionPool,
     ) -> Result<(), Error> {
         let db_client = &**get_database_client(db_pool).await?;
-        let value: JsonValue = serde_json::from_str(&self.value)?;
+        let value = match serde_json::from_str(&self.value) {
+            Ok(value) => value,
+            // `from_str` can't parse strings because they are not quoted
+            Err(_) => JsonValue::String(self.value),
+        };
         validate_editable_parameter(self.name.as_str(), &value)?;
         set_internal_property(db_client, self.name.as_str(), &value).await?;
         println!("configuration updated");