+++ mitra_models/src/properties/constants.rs
@@ -7,3 +7,4 @@ pub const ETHEREUM_CURRENT_BLOCK: &str = "ethereum_current_block";
 pub const FEDERATED_TIMELINE_RESTRICTED: &str = "federated_timeline_restricted";
 pub const FILTER_BLOCKLIST_PUBLIC: &str =  "filter_blocklist_public";
 pub const FILTER_KEYWORDS: &str = "filter_keywords";
+pub const LIKE_EMOJI: &str = "like_emoji";